refactor: atomic 구조에 맞게 폴더 변경 및 이동
