--- client/src/components/CategoryDropdown/type.ts
@@ -1,6 +1,0 @@
-interface CategoryDropdownProps {
-  state: string;
-  setState: (value: string) => void;
-}
-
-export default CategoryDropdownProps;
