--- client/src/components/QuestionTypeDropdown/type.ts
@@ -1,8 +1,0 @@
-import { QuestionType } from "types/form";
-
-interface QuestionTypeDropdownProps {
-  state: QuestionType;
-  setState: (value: QuestionType) => void;
-}
-
-export default QuestionTypeDropdownProps;
