+++ client/src/components/organisms/Edit/QuestionRead/Objective/style.ts
