+++ client/src/components/organisms/Result/QuestionResult/Checkbox/index.tsx
