+++ client/src/components/organisms/Result/QuestionResult/Multiple/index.tsx
