+++ client/src/components/organisms/Result/QuestionResult/Paragraph/index.tsx
