+++ client/src/components/organisms/Result/QuestionResult/Paragraph/style.ts
