+++ client/src/components/organisms/Result/QuestionResult/index.tsx
