+++ client/src/components/organisms/Result/QuestionResult/type.ts
