+++ client/src/components/organisms/View/QuestionView/Checkbox/style.ts
