+++ client/src/components/organisms/View/QuestionView/Paragraph/style.ts
