+++ client/src/pages/View/index.tsx
@@ -6,8 +6,8 @@ import formViewReducer from "reducer/formView";
 import formApi from "api/formApi";
 import { fromApiToForm } from "utils/form";
 import { checkPrevResponseUpdateValidateCheckList, fromApiToValidateCheckList, validationCheck } from "utils/response";
-import FormLayout from "components/Layout";
-import QuestionView from "components/QuestionView";
+import FormLayout from "components/organisms/Layout";
+import QuestionView from "components/organisms/View/QuestionView";
 import Button from "components/atoms/Button";
 import theme from "styles/theme";
 import responseApi from "api/responseApi";
